Fixed parsing bug in isBroken
isBroken parsed comments incorrectly, reacting to parens inside of comments, which it shouldn't have done. I debated fixing this one, because I'm not sure this method will be in the final version, but the fix was simple enough I slipped it in for now. Also added the default for return true.

TODO: revisit this with the idea of combining it with the actual parsing of the pgn, because then more detailed information about the error will be available.