Added test for getEnPassante
changed it into a Converter class method
It doesn't work right, so test verifies incorrect answer to prevent regression while getting the rest of the code under control. Left a TODO to revisit this method at that time to determine not only where it should reside, but how it should work.