Fixed bugs in square names and board flips
Initially, board squares were named incorrectly (black pieces were on top but top was labeled as first rank). Fixed that, and updated the tests.
Flipping board did not result in proper flipping of the square names. Updated tests to cover that, then fixed Board.flipBoard to flip the square names properly as well.