Made a note about irregularity in skipToMove.
Need to resolve it later. It currently works, but the functionality seems to depend on one error compensating for another. It goes one move to far into game, but functions using it seem to compensate by telling it to move one less. Since it works, I'm leaving it alone for now, Will revisit when the dust clears.