Fixed bug in the way comments are handled.
Comments before game started used to be thrown away. Now they are displayed.
Comments after game result used to cause program to crash. Now they are displayed.
This is the beginning of a better way to handle comments in future.

TODO:
attach comments to moves (aside from game intro and post game comment) and make that part of displaying them