Starting to build the VBoard object
The Board object in the original code was being used for too many things. There's a virtual board used many times in the code simply as an 8x8 array. I'm going to create a new object called a virtual board with only the data and methods for the board itself. Then I'll wire it into the main code body and clean out all the methods in Converter and Board that belong to it.

Also, the vBoard arrays in the current code seem inconsistent, This will unify the way the virtual board is always treated.