Created click handler for move text.
Along the way, changed the makeBold and deMakeBold methods to simply add/remove the "current_move" class to the move text. This achieved two ends: 1) it lets CSS control the appearance of the current move, and 2) it fixes a problem caused by cloning/recreating every single move element the game passes over. Jump from front to back and back in a long game, and you've created an extra 100 or more elements that aren't necessary.