Changed Image handling for board
Every square has an image, now. This makes making the moves as easy as swapping the ing src and updating the meta data.
This also prepares for a simplification of move handling, by simply updating the full position we don't need to track promotions, etc.
We might even use the same method for setting up the board as moving the pieces.
Need to update BoardImages so there's an empty image for each piece set.