Added drawFEN to Board
Since drawFEN depends on the VBoard FEN routines, VBoard was added to the demo page
drawFEN will draw either from a full FEN (with flags and indicators) or from just the position subset.
Intention is to eventually supplant all the move logic with this, because at heart a chess game is simply an ordered list of positions, the moves are merely the edges that get you from one node to another.