Added parse() the Pgn
This begins work on a new parsing algorithm. This will be a two-pass setup. The parse in Pgn will merely parse the text file into tags and linked lists of moves, containing all the information (including commentary) that can be found in the PGN file itself. The result of this parse will be fed into a second pass that will create the position maps and other information necessary to replay the games easily. By separating it like this, different game recording formats can be plugged in at the front, without impacting the game replay system.