Collapsed findFromKhight/Bishop/Rook/Queen into findFromPiece
Added better logic to detect pins and select the piece trying to move
added rank() method
Logic for selecting "from" piece now starts with a list of pieces and finds which can legally move to the destination, and disambiguates if necessary. No more move generators, except for pawns, knights and Kings.